# stylist/src/main/java/com/spinalski/stylist/api/TimeslotsApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.spinalski.stylist.api;

import java.time.OffsetDateTime;
import com.spinalski.stylist.model.TimeSlot;
import com.fasterxml.jackson.databind.ObjectMapper;
import io.swagger.annotations.*;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.servlet.http.HttpServletRequest;
import javax.validation.Valid;
import javax.validation.constraints.*;
import java.io.IOException;
import java.util.List;
import java.util.Optional;

@Api(value = "timeslots", description = "the timeslots API")
public interface TimeslotsApi {

    Logger log = LoggerFactory.getLogger(TimeslotsApi.class);

    default Optional<ObjectMapper> getObjectMapper() {
        return Optional.empty();
    }

    default Optional<HttpServletRequest> getRequest() {
        return Optional.empty();
    }

    default Optional<String> getAcceptHeader() {
        return getRequest().map(r -> r.getHeader("Accept"));
    }

    @ApiOperation(value = "get all available timeslots", nickname = "getTimeslots", notes = "", response = TimeSlot.class, responseContainer = "List", tags={ "calendar", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "all available timeslots", response = TimeSlot.class, responseContainer = "List") })
    @RequestMapping(value = "/timeslots/{startDate}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    default ResponseEntity<List<TimeSlot>> getTimeslots(@ApiParam(value = "search for timeslots starts since that timestamp",required=true) @PathVariable("startDate") OffsetDateTime startDate) {
        if(getObjectMapper().isPresent() && getAcceptHeader().isPresent()) {
            if (getAcceptHeader().get().contains("application/json")) {
                try {
                    return new ResponseEntity<>(getObjectMapper().get().readValue("[ {  \"timeSlotId\" : 0,  \"stylistId\" : 6,  \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",  \"endTime\" : \"2000-01-23T04:56:07.000+00:00\"}, {  \"timeSlotId\" : 0,  \"stylistId\" : 6,  \"startTime\" : \"2000-01-23T04:56:07.000+00:00\",  \"endTime\" : \"2000-01-23T04:56:07.000+00:00\"} ]", List.class), HttpStatus.NOT_IMPLEMENTED);
                } catch (IOException e) {
                    log.error("Couldn't serialize response for content type application/json", e);
                    return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
                }
            }
        } else {
            log.warn("ObjectMapper or HttpServletRequest not configured in default TimeslotsApi interface so no example is generated");
        }
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
    }

}